Refactoring some code into separate files
