+++ modules/backend/src/main/scala/docspell/backend/ops/OItem.scala
@@ -4,16 +4,14 @@ import cats.data.NonEmptyList
 import cats.data.OptionT
 import cats.effect.{Effect, Resource}
 import cats.implicits._
-
 import docspell.backend.JobFactory
 import docspell.common._
 import docspell.ftsclient.FtsClient
 import docspell.store.UpdateResult
-import docspell.store.queries.{QAttachment, QItem}
+import docspell.store.queries.{QAttachment, QItem, QMoveAttachment}
 import docspell.store.queue.JobQueue
 import docspell.store.records._
 import docspell.store.{AddResult, Store}
-
 import doobie.implicits._
 import org.log4s.getLogger
 
@@ -206,7 +204,7 @@ object OItem {
             target: Ident
         ): F[AddResult] =
           store
-            .transact(QItem.moveAttachmentBefore(itemId, source, target))
+            .transact(QMoveAttachment.moveAttachmentBefore(itemId, source, target))
             .attempt
             .map(AddResult.fromUpdate)
