+++ modules/joex/src/main/scala/docspell/joex/notify/NotifyDueItemsTask.scala
@@ -3,14 +3,12 @@ package docspell.joex.notify
 import cats.data.OptionT
 import cats.effect._
 import cats.implicits._
-
-import docspell.backend.ops.OItemSearch.Batch
+import docspell.backend.ops.OItemSearch.{Batch, ListItem, Query}
 import docspell.common._
 import docspell.joex.mail.EmilHeader
 import docspell.joex.scheduler.{Context, Task}
 import docspell.store.queries.QItem
 import docspell.store.records._
-
 import emil._
 import emil.builder._
 import emil.javamail.syntax._
@@ -66,11 +64,11 @@ object NotifyDueItemsTask {
       mail  <- OptionT.liftF(makeMail(sendCfg, cfg, ctx.args, items))
     } yield mail
 
-  def findItems[F[_]: Sync](ctx: Context[F, Args]): F[Vector[QItem.ListItem]] =
+  def findItems[F[_]: Sync](ctx: Context[F, Args]): F[Vector[ListItem]] =
     for {
       now <- Timestamp.current[F]
       q =
-        QItem.Query
+        Query
           .empty(ctx.args.account)
           .copy(
             states = ItemState.validStates.toList,
@@ -91,7 +89,7 @@ object NotifyDueItemsTask {
       sendCfg: MailSendConfig,
       cfg: RUserEmail,
       args: Args,
-      items: Vector[QItem.ListItem]
+      items: Vector[ListItem]
   ): F[Mail[F]] =
     Timestamp.current[F].map { now =>
       val templateCtx =
