+++ modules/store/src/main/scala/docspell/store/queries/AttachmentLight.scala
@@ -1,0 +1,10 @@
+package docspell.store.queries
+
+import docspell.common._
+
+case class AttachmentLight(
+    id: Ident,
+    position: Int,
+    name: Option[String],
+    pageCount: Option[Int]
+)
