+++ modules/store/src/main/scala/docspell/store/queries/CustomValue.scala
@@ -1,0 +1,5 @@
+package docspell.store.queries
+
+import docspell.common._
+
+case class CustomValue(field: Ident, value: String)
