+++ modules/store/src/main/scala/docspell/store/queries/ItemFieldValue.scala
@@ -1,0 +1,11 @@
+package docspell.store.queries
+
+import docspell.common._
+
+case class ItemFieldValue(
+    fieldId: Ident,
+    fieldName: Ident,
+    fieldLabel: Option[String],
+    fieldType: CustomFieldType,
+    value: String
+)
