+++ modules/store/src/main/scala/docspell/store/queries/ListItem.scala
@@ -1,0 +1,21 @@
+package docspell.store.queries
+
+import docspell.common._
+
+case class ListItem(
+    id: Ident,
+    name: String,
+    state: ItemState,
+    date: Timestamp,
+    dueDate: Option[Timestamp],
+    source: String,
+    direction: Direction,
+    created: Timestamp,
+    fileCount: Int,
+    corrOrg: Option[IdRef],
+    corrPerson: Option[IdRef],
+    concPerson: Option[IdRef],
+    concEquip: Option[IdRef],
+    folder: Option[IdRef],
+    notes: Option[String]
+)
