+++ modules/store/src/main/scala/docspell/store/queries/ListItemWithTags.scala
@@ -1,0 +1,10 @@
+package docspell.store.queries
+
+import docspell.store.records.RTag
+
+case class ListItemWithTags(
+    item: ListItem,
+    tags: List[RTag],
+    attachments: List[AttachmentLight],
+    customfields: List[ItemFieldValue]
+)
