+++ modules/store/src/main/scala/docspell/store/queries/SelectedItem.scala
@@ -1,0 +1,6 @@
+package docspell.store.queries
+
+import docspell.common._
+
+/** Some preselected item from a fulltext search. */
+case class SelectedItem(itemId: Ident, weight: Double)
